Sync dev to master

+++ python/fedml/cli/model_deployment/device_client_constants.py
@@ -51,7 +51,7 @@ class ClientConstants(object):
     MSG_MLOPS_RUN_STATUS_FAILED = "FAILED"
     MSG_MLOPS_RUN_STATUS_FINISHED = "FINISHED"
 
-    LOCAL_HOME_RUNNER_DIR_NAME = 'fedml-client'
+    LOCAL_HOME_RUNNER_DIR_NAME = 'fedml-model-client'
     LOCAL_RUNNER_INFO_DIR_NAME = 'runner_infos'
     LOCAL_PACKAGE_HOME_DIR_NAME = "fedml_packages"
 
@@ -343,74 +343,102 @@ class ClientConstants(object):
         return unzip_package_path
 
     @staticmethod
-    def cleanup_run_process():
+    def cleanup_run_process(run_id):
         try:
             local_pkg_data_dir = ClientConstants.get_data_dir()
             process_id_file = os.path.join(local_pkg_data_dir, ClientConstants.LOCAL_RUNNER_INFO_DIR_NAME,
-                                           "runner-sub-process.id")
+                                           "runner-sub-process-v2.id")
+            if not os.path.exists(process_id_file):
+                return
             process_info = load_yaml_config(process_id_file)
-            process_id = process_info.get('process_id', None)
+            if run_id is None:
+                for run_id_key, process_id_value in process_info.items():
+                    ClientConstants.cleanup_run_process(run_id_key)
+                return
+            process_id = process_info.get(str(run_id), None)
             if process_id is not None:
                 try:
                     process = psutil.Process(process_id)
                     for sub_process in process.children():
-                        os.kill(sub_process.pid, signal.SIGTERM)
+                        if platform.system() == 'Windows':
+                            os.system("taskkill /PID {} /T /F".format(sub_process.pid))
+                        else:
+                            os.kill(sub_process.pid, signal.SIGKILL)
 
                     if process is not None:
-                        os.kill(process.pid, signal.SIGTERM)
+                        if platform.system() == 'Windows':
+                            os.system("taskkill /PID {} /T /F".format(process.pid))
+                        else:
+                            os.kill(process.pid, signal.SIGKILL)
                 except Exception as e:
                     pass
-            yaml_object = {}
-            yaml_object['process_id'] = -1
-            ClientConstants.generate_yaml_doc(yaml_object, process_id_file)
+
+                process_info.pop(str(run_id))
+                ClientConstants.generate_yaml_doc(process_info, process_id_file)
         except Exception as e:
             pass
 
     @staticmethod
-    def save_run_process(process_id):
+    def save_run_process(run_id, process_id):
         try:
             local_pkg_data_dir = ClientConstants.get_data_dir()
             process_id_file = os.path.join(local_pkg_data_dir, ClientConstants.LOCAL_RUNNER_INFO_DIR_NAME,
-                                           "runner-sub-process.id")
-            yaml_object = {}
-            yaml_object['process_id'] = process_id
-            ClientConstants.generate_yaml_doc(yaml_object, process_id_file)
+                                           "runner-sub-process-v2.id")
+            if os.path.exists(process_id_file):
+                process_info = load_yaml_config(process_id_file)
+            else:
+                process_info = dict()
+            process_info[str(run_id)] = process_id
+            ClientConstants.generate_yaml_doc(process_info, process_id_file)
         except Exception as e:
             pass
 
     @staticmethod
-    def cleanup_learning_process():
+    def cleanup_learning_process(run_id):
         try:
             local_pkg_data_dir = ClientConstants.get_data_dir()
             process_id_file = os.path.join(local_pkg_data_dir, ClientConstants.LOCAL_RUNNER_INFO_DIR_NAME,
-                                           "runner-learning-process.id")
+                                           "runner-learning-process-v2.id")
             process_info = load_yaml_config(process_id_file)
-            process_id = process_info.get('process_id', None)
+            if run_id is None:
+                for run_id_key, process_id_value in process_info.items():
+                    ClientConstants.cleanup_learning_process(run_id_key)
+                return
+            process_id = process_info.get(str(run_id), None)
             if process_id is not None:
                 try:
                     process = psutil.Process(process_id)
                     for sub_process in process.children():
-                        os.kill(sub_process.pid, signal.SIGTERM)
+                        if platform.system() == 'Windows':
+                            os.system("taskkill /PID {} /T /F".format(sub_process.pid))
+                        else:
+                            os.kill(sub_process.pid, signal.SIGTERM)
 
                     if process is not None:
-                        os.kill(process.pid, signal.SIGTERM)
+                        if platform.system() == 'Windows':
+                            os.system("taskkill /PID {} /T /F".format(process.pid))
+                        else:
+                            os.kill(process.pid, signal.SIGTERM)
                 except Exception as e:
                     pass
-            yaml_object = {}
-            yaml_object['process_id'] = -1
-            ClientConstants.generate_yaml_doc(yaml_object, process_id_file)
+
+                process_info.pop(str(run_id))
+                ClientConstants.generate_yaml_doc(process_info, process_id_file)
         except Exception as e:
             pass
 
     @staticmethod
-    def save_learning_process(learning_id):
+    def save_learning_process(run_id, learning_id):
         try:
             local_pkg_data_dir = ClientConstants.get_data_dir()
             process_id_file = os.path.join(local_pkg_data_dir, ClientConstants.LOCAL_RUNNER_INFO_DIR_NAME,
-                                           "runner-learning-process.id")
-            yaml_object = {}
-            yaml_object['process_id'] = learning_id
-            ClientConstants.generate_yaml_doc(yaml_object, process_id_file)
+                                           "runner-learning-process-v2.id")
+            if os.path.exists(process_id_file):
+                process_info = load_yaml_config(process_id_file)
+            else:
+                process_info = dict()
+            process_info[str(run_id)] = learning_id
+            ClientConstants.generate_yaml_doc(process_info, process_id_file)
         except Exception as e:
             pass
 
@@ -501,14 +529,16 @@ class ClientConstants(object):
 
         if platform.system() == 'Windows':
             if no_sys_out_err:
-                script_process = subprocess.Popen(script_path)
+                script_process = subprocess.Popen(script_path, preexec_fn=os.setsid)
             else:
-                script_process = subprocess.Popen(script_path, stdout=stdout_flag, stderr=stderr_flag)
+                script_process = subprocess.Popen(script_path, stdout=stdout_flag, stderr=stderr_flag,
+                                                  preexec_fn=os.setsid)
         else:
             if no_sys_out_err:
-                script_process = subprocess.Popen(['bash', '-c', script_path])
+                script_process = subprocess.Popen(['bash', '-c', script_path], preexec_fn=os.setsid)
             else:
-                script_process = subprocess.Popen(['bash', '-c', script_path], stdout=stdout_flag, stderr=stderr_flag)
+                script_process = subprocess.Popen(['bash', '-c', script_path], stdout=stdout_flag, stderr=stderr_flag,
+                                                  preexec_fn=os.setsid)
 
         return script_process
 
@@ -517,7 +547,8 @@ class ClientConstants(object):
         stdout_flag = subprocess.PIPE if should_capture_stdout else sys.stdout
         stderr_flag = subprocess.PIPE if should_capture_stderr else sys.stderr
 
-        script_process = subprocess.Popen([shell, script_path], stdout=stdout_flag, stderr=stderr_flag)
+        script_process = subprocess.Popen([shell, script_path], stdout=stdout_flag, stderr=stderr_flag,
+                                          preexec_fn=os.setsid)
 
         return script_process
 
@@ -527,7 +558,8 @@ class ClientConstants(object):
         stdout_flag = subprocess.PIPE if should_capture_stdout else sys.stdout
         stderr_flag = subprocess.PIPE if should_capture_stderr else sys.stderr
 
-        script_process = subprocess.Popen(shell_script_list, stdout=stdout_flag, stderr=stderr_flag)
+        script_process = subprocess.Popen(shell_script_list, stdout=stdout_flag, stderr=stderr_flag,
+                                          preexec_fn=os.setsid)
 
         return script_process
